added loading, signed, submitted, completed states for each tx that user confirms.
