+++ Sources/BraveWallet/Assets.xcassets/Brave Wallet/Transaction States/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
